requirements.txt
django-allauth 
dj-rest-auth
 google-auth
  requests
  python-dotenv

 google-auth
  google-auth-oauthlib
   google-auth-httplib2
    google-auth-oauthlib
    channels
    daphne


